+++ common/db/src/main/java/is/codion/common/db/database/package-info.java
@@ -21,6 +21,7 @@
  * <p>Package configuration values:
  * <ul>
  * <li>{@link is.codion.common.db.database.Database#DATABASE_URL}
+ * <li>{@link is.codion.common.db.database.Database#URL_SCOPED_INSTANCE}
  * <li>{@link is.codion.common.db.database.Database#DATABASE_INIT_SCRIPTS}
  * <li>{@link is.codion.common.db.database.Database#COUNT_QUERIES}
  * <li>{@link is.codion.common.db.database.Database#SELECT_FOR_UPDATE_NOWAIT}